Added better bottom bar handling, fixes to esl_central

+++ firmware/common/include/esl_packets.h
@@ -12,8 +12,8 @@ struct esl_coordinate {
 } __packed;
 
 struct esl_sensor_reading {
-    uint16_t temperature;
-    uint16_t humidity;
+    float temperature;
+    float humidity;
 } __packed;
 
 #endif
